requirements.txt
fastapi
uvicorn
sentence-transformers==2.2.2
faiss-cpu
transformers==4.28.1
huggingface-hub==0.14.1
requests
